Renaming loading state component.
