+++ src/app/image-group-page/image-group-page.component.pug
@@ -14,7 +14,7 @@ div
           | Unable to load the thumbnails for this group.
 
 ang-google-auth(*ngIf="showGoogleAuth", (closeModal)="closeGoogleAuth($event)")
-ang-export-loading-state(*ngIf="showExportLoadingState", [options]="exportLoadingStateopts", (closeModal)="closeExportLoadingState($event)")
+ang-loading-state(*ngIf="showExportLoadingState", [options]="exportLoadingStateopts", (closeModal)="closeExportLoadingState($event)")
 ang-terms-and-conditions(*ngIf="showTermsConditions", [exportType]="exportType", (closeModal)="handleTCModalClose($event)", [ig]="ig")
 ang-ig-download-modal(*ngIf="showPptModal", (closeModal)="showPptModal = false", [ig]="ig")
 ang-download-limit-modal(*ngIf="showDownloadLimitModal", (closeModal)="showDownloadLimitModal = false", [ig]="ig")
