+++ src/app/modals/index.ts
@@ -17,4 +17,4 @@ export * from './edit-personal-collection/edit-personal-collection.component'
 export * from './register-jstor-modal/register-jstor-modal.component'
 export * from './change-password/change-password.component'
 export * from './google-auth/google-auth.component'
-export * from './export-loading-state/export-loading-state.component'
+export * from './loading-state/loading-state.component'
