+++ src/app/modals/loading-state/loading-state.component.scss
